backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
fastapi
uvicorn[standard]
motor
pydantic
python-jose
passlib[bcrypt]
flask
flask-cors
pydantic[email]